+++ pg_lakehouse/src/duckdb/mod.rs
@@ -1,1 +1,2 @@
 pub mod connection;
+pub mod parquet;
